Next iteration on DO ... CONTINUE
Turns out the examples in the issues were too simple. CONTINUE has a
number of uses that are hard or impossible to catch.
This commit makes sure the LABEL is at the start of the line. Does
handle some cases better (also better than master), but breaks a lot of
other cases. Perhaps this construct it simply too old.